Add the checking for incorrect input in edit lines
For the working with alignment must use to VLayout which was defined in ui file. It contains fields like labels, edit lines and buttons.
If paths are incorrect, then write an error message about it. In order to do it have to insert the error message after the corresponding edit line. So, it's necessary to know indices of edit lines. After the writing message, increment the index of second edit lines, because it shifts due to was inserted the error message. Also, the accept button must be disabled if at least one of the directories is incorrect.
Incorrect_maps(troops)_directory is necessary flag in order to write the error message after the line of edit if it's incorrect at the moment.