Add an overwrite function for mouse's wheel scrolling, rewrite the function for setting the map
It's important to process keystrokes on the keyboard and save it to the set. Built-in functions don't provide to handle keyboard state in PyQt. So, all keystrokes will be saved into the set, then they can be handled.

Instead of QLabel and QPixmap I had to use QGraphicsScene which provides easier scrolling on the map or any scaling on it.