Delete outdated values zoom_in/out, specify new parameters for scaling the view
The current scale is necessary for determining the current scale of view, this value will change depending on the scale of view.
The max scale is necessary to set limit on maximum scaling.
The min scale was determined for 2 cases: for scaling based on width and for scaling based on height of the available area on the window intended for setting the map.
Scale factor specifies a value by which the scale's value of view will change.