Remove from class unnecessary slots, edit method of setting menu bar
Menu bar is set by particular class for working with it. Therefore, the method for setting the menu bar was edited.
Also, a method of setting a central widget was be added. It's a slot, that is called when all settings are correct and the accept button is pressed in the dialog window of creating new map.